requirements.txt
altair==4.1.0
appdirs==1.4.3
astor==0.8.1
astroid==2.4.2
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
boto3==1.14.27
botocore==1.17.27
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.0
docutils==0.15.2
emoji==0.5.4
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
flasgger==0.9.4
Flask==1.1.1
Flask-Cors==3.0.8
gunicorn==19.9.0
idna==2.10
imutils==0.5.3
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.1
opencv-python==4.3.0.36
packaging==20.4
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
pyarrow==0.17.1
pydeck==0.4.0
Pygments==2.6.1
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.14.0
streamlit==0.64.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.10
validators==0.16.0
virtualenv==20.0.7
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.15.5
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xgboost==1.1.1
